+++ src/domain/services/user-items/get-user-items.ts
@@ -2,7 +2,7 @@ import { selectUserItems } from '@/db/repositories/user-item-repository'
 import type { SelectUserItems } from '@/domain/entities/user-item'
 
 export async function getUserItemsService(input: SelectUserItems) {
-  const userItems = await selectUserItems(input)
+  const userItems = await selectUserItems({ ...input })
 
   const lastUserItem = userItems[input.pageSize]
   const nextCursor = lastUserItem?.updatedAt.toISOString() || null